refactor: empty work/ directory and fix admin handler imports
Moved status.ts to commands/status.ts, moved hook/tldr/linear-states
handlers to admin/ with correct lib import depths. Deleted legacy
work/sync.ts (shadow sync not in new taxonomy). work/ directory is
now empty.
